fix: rename Contests model to contests
